# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de Atendimento",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/clientes/find/{cliente}": {
            "get": {
                "tags": [
                    "/clientes/find/:cliente"
                ],
                "summary": "Pega cliente por Id",
                "description": "Retorna cliente por Id",
                "operationId": "App\\Http\\Controllers\\ClienteController::show",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "Cliente id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/clientes/create": {
            "post": {
                "tags": [
                    "/cliente/create"
                ],
                "summary": "Cria um novo Cliente",
                "description": "Retorna um novo cliente",
                "operationId": "App\\Http\\Controllers\\ClienteController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/clientes/update/{cliente}": {
            "put": {
                "tags": [
                    "/cliente/update/:cliente"
                ],
                "summary": "Atualiza cliente existente",
                "description": "Retorna cliente atualizado por id",
                "operationId": "App\\Http\\Controllers\\ClienteController::update",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "Cliente id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Sucesso"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/clientes/{cliente}": {
            "delete": {
                "tags": [
                    "/clientes/:cliente"
                ],
                "summary": "Deleta um cliente existente",
                "description": "Deleta o cliente por Id",
                "operationId": "App\\Http\\Controllers\\ClienteController::delete",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "Cliente id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/clientes/addServicos/{cliente}": {
            "post": {
                "tags": [
                    "/clientes/addServicos/:cliente"
                ],
                "summary": "Adiciona servicos ao cliente",
                "description": "Adiciona servicos ao cliente",
                "operationId": "App\\Http\\Controllers\\ClienteController::storeServicos",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "cliente id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "servicos": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}